requirements.txt
cassandra-driver==3.29.0
pymongo==4.6.0
elasticsearch==8.13.0
